refactor: move Release modal into footer component
