+++ frontend/src/App.vue
@@ -22,73 +22,38 @@
         <Footer />
       </div>
     </div>
-
-    <!-- Release Notification Modal at App level -->
-    <ReleaseNotificationModal
-      modal-id="globalReleaseModal"
-      :release="releasesStore.latestRelease"
-      @dismiss="handleReleaseDissmiss"
-    />
   </div>
 </template>
 
 <script setup>
-import { computed, onMounted } from 'vue'
-import { RouterView, useRoute } from 'vue-router'
+import { computed, watch } from 'vue'
+import { RouterView, useRoute, useRouter } from 'vue-router'
 import { useNotificationsStore } from '@/store/notifications.js'
-import { useReleasesStore } from '@/store/releases.js'
-import { useActuatorInfoStore } from '@/store/actuatorInfo.js'
 import allFeatureFlags from '@/config/featureFlags.js'
 import VueTour from './components/VueTour.vue'
 import NavigationBar from './components/NavigationBar.vue'
 import AlertNotification from './components/AlertNotification.vue'
 import Footer from './components/FooterComp.vue'
 import errorPage from '@/views/ErrorPage.vue'
-import ReleaseNotificationModal from './components/modals/ReleaseNotificationModal.vue'
 import { useUiConfiguration } from '@/store/uiConfiguration.js'
 
 const uiConfigurationStore = useUiConfiguration()
 const useNotifications = useNotificationsStore()
-const releasesStore = useReleasesStore()
-const actuatorInfoStore = useActuatorInfoStore()
 const route = useRoute()
+const router = useRouter()
 
 const vueTourFeatureFlag = !!(
   allFeatureFlags.vueTour === 'true' || allFeatureFlags.vueTour === true
 )
 
-// Initialize release checking system at app level
-onMounted(() => {
-  releasesStore.loadDismissedReleases()
-
-  // Set current version when actuator info is available
-  if (actuatorInfoStore.actuatorInfoBuildVersion) {
-    releasesStore.setCurrentVersion(actuatorInfoStore.actuatorInfoBuildVersion)
-    // Only start periodic checks if version pattern is supported
-    if (isVersionSupported(actuatorInfoStore.actuatorInfoBuildVersion)) {
-      releasesStore.startPeriodicCheck()
+watch(
+  () => router.currentRoute.value.fullPath,
+  (newVal, oldVal) => {
+    if (oldVal !== '/') {
+      useNotifications.criticalError = false
     }
-  } else {
-    // Wait for actuator info to be loaded
-    actuatorInfoStore.retrieveBackendActuatorInfo().then(() => {
-      releasesStore.setCurrentVersion(actuatorInfoStore.actuatorInfoBuildVersion)
-      // Only start periodic checks if version pattern is supported
-      if (isVersionSupported(actuatorInfoStore.actuatorInfoBuildVersion)) {
-        releasesStore.startPeriodicCheck()
-      }
-    })
-  }
-})
-
-function isVersionSupported(version) {
-  // Check if version follows v3.x.x pattern (e.g., v3.17.3, v3.0.0, v3.1.2)
-  const versionPattern = /^v3\.\d+\.\d+$/
-  return versionPattern.test(version)
-}
-
-function handleReleaseDissmiss(tagName) {
-  releasesStore.dismissRelease(tagName)
-}
+  },
+)
 
 const isVueTourVisible = computed(() => {
   return (
